example/gharchive/requirements.txt
sqlalchemy
coloredlogs
pygithub